+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Most recent version is listed first.
 # v0.0.49
 - Add mux Resolve function: https://github.com/komuw/ong/pull/268
 - Use http.Handler as the http middleware instead of http.HandlerFunc: https://github.com/komuw/ong/pull/269
+- Add optional http timeout: https://github.com/komuw/ong/pull/270
 
 ## v0.0.48
 - Change attest import path: https://github.com/komuw/ong/pull/265